requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
python-multipart==0.0.6
tensorflow-cpu==2.13.0
numpy==1.24.3
Pillow==10.0.0
h5py==3.9.0